Implement validations and basic escape game creation
Links still need validating, and I could do with tightening validation
on other things.

Enums are numbered hashes for the sake of flexibility. If genres are
removed or added, this can be done without retroactively creating
conflicts.

Validations are loose on Google Maps Place IDs because even Google
hasn't defined the length of these very well. I'll need to do more
validation in the frontend.