Limit number of images to be uploaded in frontend
This also helped me get to the source of #89. InvalidSignature is raised
when an invalid signed_id (signature) for an image is passed. These
linger when an image isn't uploaded, but is still kept on the form, so
the solution is to only account for images with an id (those that have
been uploaded).